Improved: Bootstrap version upgraded to 5.3.3 Improved: Code has been improved to support PDUs with two-digit outlets. <a href="https://github.com/disisto/apc-switched-rack-pdu-control-panel/pull/5/commits">PR #5</a>
⚠️ This is the last iteration of the 1.x versions.

With a focus on maintainability, usability, performance, and security, the principle of the monolithic/portable script is being abandoned. Version 2.x, which has been completely rebuilt from the ground up and incorporates feedback from the past 12 months, will be released shortly.

If you have not yet integrated the script into your environment, it is advisable to wait with the adjustments.